contribution을 matrix로 만드는 함수를 별도로 분리하고 useMemo를 쓴다
